Move Repo tto a component
