+++ src/components/index.js
@@ -9,3 +9,4 @@ export { default as ErrorMessage } from './ErrorMessage'
 export { default as Loader } from './Loader'
 export { default as DataToShow } from './DataToShow'
 export { default as ReposList } from './ReposList'
+export { default as Repo } from './Repo'
